Services refactor
Refactors the server directory to more explicitly state that the modules
contained within are actually services. Also attempts to introduce more
generic directory structures for each service:

- renames `src/servers` to `src/services`
- renames `src` directories within the service module to `service`
- includes an `nginx.conf` file in every service directory
- includes a `README.md` for every service

closes #177
