--- src/servers/spotlight/conf.mjs
@@ -1,4 +1,0 @@
-import * as path from 'path';
-
-export const SERVER_DIRECTORY = 'src/servers/spotlight';
-export const TERRAFORM_DIRECTORY = path.join(SERVER_DIRECTORY, 'terraform');
