--- src/servers/spotlight/terraform/README.md
@@ -1,4 +1,0 @@
-# Spotlight Infrastructure as Code
-
-This directory is used to perform commands with Terraform to provision EC2
-instances which can then be used for efficient Spotlight Annotation.
