+++ src/services/annotation/README.md
@@ -1,0 +1,3 @@
+# Annotation Service
+
+Documentation is left as a TODO
