+++ src/services/annotation/service/app.mjs
@@ -1,11 +1,10 @@
 import Fastify from 'fastify';
 
-import { PORT } from './config.mjs';
+import { PORT } from '../config.mjs';
 import { routes } from './routes.mjs';
-import { __dirname } from './util.mjs';
 
 const fastify = Fastify({
-	logger: false
+	logger: true
 });
 
 fastify.register(routes);
@@ -13,9 +12,11 @@ fastify.register(routes);
 const start = async () => {
 	try {
 		await fastify.listen({ port: PORT });
+		console.log(`Listening at http://localhost:${PORT}`);
 	} catch (err) {
 		fastify.log.error(err);
 		throw new Error(err);
 	}
 };
+
 start();
