+++ src/services/annotation/service/machine.mjs
@@ -3,6 +3,7 @@ import { createMachine, assign, interpret } from 'xstate';
 import { from, map, mergeAll } from 'rxjs';
 
 import * as spotlight from 'dbpedia/spotlight.mjs';
+
 import * as terraform from './terraform.mjs';
