+++ src/services/evaluator/nginx.conf
@@ -1,0 +1,8 @@
+events {}
+http {
+    server {
+        location / {
+            proxy_pass http://localhost:4000;
+        }
+    }
+}
