+++ src/services/spotlight/nginx.conf
