+++ src/services/spotlight/service/state.mjs
