+++ src/services/spotlight/spotlightDockerCommand.sh
